requirement.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.24.85
botocore==1.27.85
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==37.0.4
Django==4.1
django-cors-headers==3.13.0
django-filter==22.1
django-storages==1.13.1
djangorestframework==3.13.1
idna==3.3
jmespath==1.0.1
Pillow==9.2.0
pycparser==2.21
PyJWT==2.5.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.2.1
razorpay==1.3.0
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
twilio==7.12.1
urllib3==1.26.11